Revert "Fix text wrapping in commit message editors (#31030)"
This reverts commit f2601ce52ce82eb201799ae6c4f1f92f42ccf7c8.
