+++ assets/settings/default.json
@@ -1452,9 +1452,7 @@
       "language_servers": ["erlang-ls", "!elp", "..."]
     },
     "Git Commit": {
-      "allow_rewrap": "anywhere",
-      "preferred_line_length": 72,
-      "soft_wrap": "bounded"
+      "allow_rewrap": "anywhere"
     },
     "Go": {
       "code_actions_on_format": {
